Add asset config to browser builder
