+++ src/builders/browser/browser-transforms.ts
@@ -7,15 +7,25 @@ import {
 } from 'webpack';
 
 import {
+  applyAppAssetsConfig
+} from '../../webpack/app-assets-utils';
+
+import {
   SkyuxSaveHostMetadataPlugin
 } from '../../webpack/plugins/save-host-metadata/save-host-metadata-plugin';
+
+import {
+  SkyuxBrowserBuilderOptions
+} from './browser-options';
 
 /**
  * Allows adjustments to the default Angular "browser" webpack config.
  * @param options The input options passed to the builder.
  * @param context The context of the builder execution.
  */
-function getBrowserWepbackConfigTransformer(): ExecutionTransformer<WebpackConfig> {
+function getBrowserWepbackConfigTransformer(
+  options: SkyuxBrowserBuilderOptions
+): ExecutionTransformer<WebpackConfig> {
   return (webpackConfig) => {
 
     webpackConfig.plugins = webpackConfig.plugins || [];
@@ -24,13 +34,15 @@ function getBrowserWepbackConfigTransformer(): ExecutionTransformer<WebpackConfi
       new SkyuxSaveHostMetadataPlugin()
     );
 
+    applyAppAssetsConfig(webpackConfig, options);
+
     return webpackConfig;
   };
 
 }
 
-export function getBrowserTransforms() {
+export function getBrowserTransforms(options: SkyuxBrowserBuilderOptions) {
   return {
-    webpackConfiguration: getBrowserWepbackConfigTransformer()
+    webpackConfiguration: getBrowserWepbackConfigTransformer(options)
   };
 }
