+++ src/builders/dev-server/dev-server-transforms.ts
@@ -7,20 +7,12 @@ import {
 } from '@angular-devkit/build-angular';
 
 import {
-  AngularCompilerPlugin
-} from '@ngtools/webpack';
-
-import {
   Configuration as WebpackConfig
 } from 'webpack';
 
 import {
-  getAppAssetsRules
-} from '../../webpack/loaders/app-assets/app-assets-rules';
-
-import {
-  SkyuxAppAssetsPlugin
-} from '../../webpack/plugins/app-assets/app-assets.plugin';
+  applyAppAssetsConfig
+} from '../../webpack/app-assets-utils';
 
 import {
   SkyuxOpenHostURLPlugin
@@ -35,15 +27,6 @@ import {
 } from './dev-server-utils';
 
 /**
- * Allows other Webpack loaders to process component HTML templates.
- * @see https://github.com/angular/angular-cli/issues/15861
- */
-function unlockComponentTemplates(webpackConfig: WebpackConfig): void {
-  const compilerPlugin = webpackConfig.plugins?.find(plugin => plugin instanceof AngularCompilerPlugin);
-  (compilerPlugin as AngularCompilerPlugin).options.directTemplateLoading = false;
-}
-
-/**
  * Allows adjustments to the default Angular "dev-server" webpack config.
  * @param options The input options passed to the builder.
  * @param context The context of the builder execution.
@@ -55,12 +38,10 @@ function getDevServerWepbackConfigTransformer(
   return (webpackConfig) => {
 
     webpackConfig.plugins = webpackConfig.plugins || [];
-    webpackConfig.module = webpackConfig.module || { rules: [] };
 
     if (options.skyuxLaunch === 'host') {
       /*istanbul ignore next line*/
       const pathName = context.target?.project!;
-      const assetBaseUrl = options.deployUrl!;
 
       webpackConfig.plugins.push(
         new SkyuxOpenHostURLPlugin({
@@ -70,15 +51,7 @@ function getDevServerWepbackConfigTransformer(
         })
       );
 
-      webpackConfig.plugins.push(
-        new SkyuxAppAssetsPlugin()
-      );
-
-      webpackConfig.module.rules.push(
-        ...getAppAssetsRules(assetBaseUrl)
-      );
-
-      unlockComponentTemplates(webpackConfig);
+      applyAppAssetsConfig(webpackConfig, options);
     }
 
     return webpackConfig;
